somd docs, some test
