+++ src/main/java/Server/Rule_SBLeft.java
@@ -1,10 +1,19 @@
 package Server;
 
+/**
+ * if sb left game say to everybody
+ */
 public class Rule_SBLeft extends rule {
     Rule_SBLeft(serverHead head) {
         super(head);
     }
 
+    /**
+     * if sb left game say to everybody
+     * @param player who done action
+     * @param command what action
+     * @return true if sb left
+     */
     @Override
     boolean check(player player, String command) {
         System.out.println("Try " + this.getClass().getSimpleName());
