+++ src/main/java/Server/Rule_WaitForAll.java
@@ -1,11 +1,20 @@
 package Server;
 
+/**
+ * check is all players to play
+ */
 public class Rule_WaitForAll extends rule {
 
     Rule_WaitForAll(serverHead head) {
         super(head);
     }
 
+    /**
+     * check is all players to play
+     * @param player who done action
+     * @param command what action
+     * @return true if there not enought players
+     */
     @Override
     boolean check(player player, String command) {
         System.out.println("Try " + this.getClass().getSimpleName());
