+++ src/main/java/Server/logicBoard.java
@@ -1,10 +1,18 @@
 package Server;
 
+/**
+ * servers board for game
+ */
 public abstract class logicBoard {
 
     field[][] fields;
 
-
+    /**
+     * decode fast way to sets fields on board
+     * @param createTables set of code, have fields
+     * @param color whats color to paint
+     * @param winningHause code field winningColor
+     */
     protected void decoder(int[] createTables, String color, String winningHause) {
 
             if (createTables[0] == 1) {
@@ -25,6 +33,13 @@ public abstract class logicBoard {
             }
     }
 
+    /**
+     * to check all winning fields are captured
+     * @param createdTables
+     * @param color
+     * @param positiveHouse
+     * @return
+     */
     protected boolean checker(int[] createdTables, String color, int positiveHouse) {
 
         int counter = 0;
@@ -55,13 +70,35 @@ public abstract class logicBoard {
         }
     }
 
+    /**
+     * fields color getter
+     * @param x fields x
+     * @param y fields y
+     * @return fields x,y's color
+     */
     public String getFieldColor(int x, int y){
         return fields[x][y].getColor();
     }
+
+    /**
+     * fields color setter
+     * @param x fields x
+     * @param y fields y
+     * @param color new color
+     */
     public void setFieldColor(int x, int y,String color) {
         fields[x][y].setColor(color);
     }
+
+    /**
+     * @return whats winning color for field x,y
+     *
+     */
     public String getFieldHause(int x, int y) { return fields[x][y].colorToWin; }
 
+    /**
+     * check are all winning fields caputered by sb
+     * @return fact
+     */
     protected abstract String hasSbWon();
 }
