+++ src/main/java/Server/player.java
@@ -6,6 +6,9 @@ import java.io.PrintWriter;
 import java.net.Socket;
 import java.util.Scanner;
 
+/**
+ * all info about players, sockets start, communication etc
+ */
 public class player implements Runnable {
     Socket socket;
     String color;
@@ -20,6 +23,9 @@ public class player implements Runnable {
         this.head = head;
     }
 
+    /**
+     * when client starts exitng done this
+     */
     @Override
     public void run() {
         try {
@@ -36,12 +42,19 @@ public class player implements Runnable {
         }
     }
 
+    /**
+     * cannals to communicat and say hello
+     * @throws IOException
+     */
     private void setup() throws IOException {
         input = new Scanner(socket.getInputStream());
         output = new PrintWriter(socket.getOutputStream(), true);
         output.println("WELCOME " + color);
     }
 
+    /**
+     * always wait for new message
+     */
     private void listenClient() {
         while (input.hasNextLine()) {
             var command = input.nextLine();
@@ -49,10 +62,18 @@ public class player implements Runnable {
         }
     }
 
+    /**
+     * send message to client whom is a player
+     * @param message message to deliver
+     */
     public void sendMessage(String message){
         output.println(message);
     }
 
+    /**
+     * players color getter
+     * @return players color
+     */
     public String getColor() {
         return color;
     }
