+++ src/main/java/Server/serverExec.java
@@ -1,6 +1,13 @@
 package Server;
 
+/**
+ * class to start server
+ */
 public class serverExec {
+    /**
+     * main to start server
+     * @param args useles
+     */
     public static void main(String[] args){
         setUpServer.openMenuWindow();
     }
