+++ src/main/java/Server/setUpServer.java
@@ -2,9 +2,12 @@ package Server;
 
 import java.util.ArrayList;
 
+/**
+ * allow to set game setting, prepare settings and open server for game
+ */
 public class setUpServer {
 
-	private static serverHead game = new serverHead();;
+	private static serverHead game = new serverHead();
 	public static void createNewGame(logicBoard board, int howManyPlayers, String shape, ArrayList<rule> listRule, ArrayList<RuleMove> listMove){
 
 		for(RuleMove ruleMove: listMove){
